Moved AnimatorParameters to animation namespace, added triggering death animation
