+++ Assets/Scripts/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85672c33380de1f41a16501a8438b0c3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
