+++ Assets/Scripts/Combat/Fighter.cs
@@ -1,6 +1,7 @@
 using UnityEngine;
 using RPG.Movement;
 using RPG.Core;
+using RPG.Animation;
 
 namespace RPG.Combat
 {
